Add eval type selection based on rank
