+++ src/store/evalFormFields.js
@@ -66,3 +66,73 @@ export const NAVPERS_1616_26_REV11_11 = {
   "form1[0].Page2[0]._49_SrRater_Sign_Date[0]": "",
   "form1[0].Page2[0].fontsize[0]": ""
 };
+
+export const NAVPERS_1616_27_REV08_10 = {
+  "form1[0].Page1[0]._32_Ind_Sig_Counseled[0]": "",
+  "form1[0].Page1[0]._5_Status[0]": "",
+  "form1[0].Page1[0]._10_Periodic[0]": "",
+  "form1[0].Page1[0]._11_Detach_ind[0]": "",
+  "form1[0].Page1[0]._12_Detach_Sr[0]": "",
+  "form1[0].Page1[0]._13_Special[0]": "",
+  "form1[0].Page1[0]._16_NOB[0]": "",
+  "form1[0].Page1[0]._17_Report_Regular[0]": "",
+  "form1[0].Page1[0]._18_Report_Concurrent[0]": "",
+  "form1[0].Page1[0]._19_Report_OpsCDR[0]": "",
+  "form1[0].Page1[0]._31_Counselor[0]": "",
+  "form1[0].Page1[0]._33_Leadership[0]": "",
+  "form1[0].Page1[0]._34_InstExpertise[0]": "",
+  "form1[0].Page1[0]._35_Professionalism[0]": "",
+  "form1[0].Page1[0]._36_Loyalty[0]": "",
+  "form1[0].Page1[0]._20_Physical_Readiness[0]": "",
+  "form1[0].Page1[0]._21_Billet[0]": "",
+  "form1[0].Page1[0]._22_Reporting_Senior_Last__FI_MI[0]": "",
+  "form1[0].Page1[0]._24_Desig_Rpt_Sr[0]": "",
+  "form1[0].Page1[0]._25_Title_Rpt_Sr[0]": "",
+  "form1[0].Page1[0]._26_UIC_Rpt_Sr[0]": "",
+  "form1[0].Page1[0]._27_SSN_Rpt_Sr[0]": "",
+  "form1[0].Page1[0]._28_Command_achievements[0]": "",
+  "form1[0].Page1[0]._29_PrimaryCollat[0]": "",
+  "form1[0].Page1[0]._23_Grade_Rpt_Sr[0]": "",
+  "form1[0].Page1[0]._1_Name_Last__First_MI_Suffix[0]": "",
+  "form1[0].Page1[0]._2_GradeRate[0]": "",
+  "form1[0].Page1[0]._3_Desig[0]": "",
+  "form1[0].Page1[0]._4_SSN[0]": "",
+  "form1[0].Page1[0]._6_UIC[0]": "",
+  "form1[0].Page1[0]._7_ShipStation[0]": "",
+  "form1[0].Page1[0]._8_Promotion_Status[0]": "",
+  "form1[0].Page1[0]._9_Date_Reported[0]": "",
+  "form1[0].Page1[0]._30_Date_Counsel[0]": "",
+  "form1[0].Page1[0]._14_PeriodDate_From[0]": "",
+  "form1[0].Page1[0]._15_PeriodDate_To[0]": "",
+  "form1[0].Page1[0].Text[0].ResetButton1[0]": "",
+  "form1[0].Page1[0].Text[0].SaveButton[0]": "",
+  "form1[0].Page1[0].Lines[0]._29_PrimaryCollat_Short[0]": "",
+  "form1[0].Page2[0]._42_Recommendation[0]": "",
+  "form1[0].Page2[0]._37_Character[0]": "",
+  "form1[0].Page2[0]._38_Communication[0]": "",
+  "form1[0].Page2[0]._39_Heritage[0]": "",
+  "form1[0].Page2[0]._1_Name_Last__First_MI_Suffix[0]": "",
+  "form1[0].Page2[0]._2_GradeRate[0]": "",
+  "form1[0].Page2[0]._3_Desig[0]": "",
+  "form1[0].Page2[0]._4_SSN[0]": "",
+  "form1[0].Page2[0]._40_Recommend1[0]": "",
+  "form1[0].Page2[0]._40_Recommend2[0]": "",
+  "form1[0].Page2[0]._41_Comments_on_Perf[0]": "",
+  "form1[0].Page2[0]._43_Summary_Sig_Problem[0]": "",
+  "form1[0].Page2[0]._43_Summary_Progressing[0]": "",
+  "form1[0].Page2[0]._43_Summary_Promotable[0]": "",
+  "form1[0].Page2[0]._43_Summary_Must_Promote[0]": "",
+  "form1[0].Page2[0]._43_Summary_Early_Promote[0]": "",
+  "form1[0].Page2[0]._44_Rpt_Sr_Address[0]": "",
+  "form1[0].Page2[0].Member_Trait_Avg[0]": "",
+  "form1[0].Page2[0]._45_Rpt_Sr_Signature[0]": "",
+  "form1[0].Page2[0]._45_Rpt_Sr_Sign_Date[0]": "",
+  "form1[0].Page2[0]._46_Date[0]": "",
+  "form1[0].Page2[0]._46_Submit[0]": "",
+  "form1[0].Page2[0]._46_Signature[0]": "",
+  "form1[0].Page2[0]._47_Reg_Rpt_Sr_Sig[0]": "",
+  "form1[0].Page2[0]._47_Date[0]": "",
+  "form1[0].Page2[0].Lines[0]._46_No_Submit[0]": "",
+  "form1[0].Page2[0].Lines[0].fontsize[0]": "",
+  "form1[0]": ""
+};
