+++ src/store/index.js
@@ -29,9 +29,7 @@ export default new Vuex.Store({
     errorMsg: state => state.app.errorMsg,
     memberStatuses: state => state.defaults.memberStatuses,
     promotionStatuses: state => state.defaults.promotionStatuses,
-    ranksAll: state => state.defaults.ranks,
-    ranksEnlisted: state => state.defaults.ranks.enlisted,
-    ranksOfficer: state => state.defaults.ranks.officer,
+    ranks: state => state.defaults.ranks,
     traits: state => state.defaults.traits,
     promotionRecommendations: state => state.defaults.promotionRecommendations,
     fonts: state => state.defaults.fonts,
